# openapi.yaml
openapi: 3.0.3
info:
  title: API MongoDB — Lista de Exercícios
  version: 1.0.0
  description: |-
    Lista de Exercícios de MongoDB implementada como API REST com **Node.js + Express**
    e o driver oficial `mongodb` (sem ODM).

    Cada grupo de rotas corresponde a uma parte da lista:

    | Recurso | Exercícios | Conceitos |
    |---|---|---|
    | Produtos | 2 a 11 | CRUD, `$gt`/`$lt`, `$and`/`$or`/`$nor`, projeção, `sort`/`limit`, `$set`, `$inc` |
    | Clientes | 12, 13, 14 | documentos aninhados, `$push`, `$pull`, `$addToSet` |
    | Filmes | 15 | modelagem com array + documento aninhado |
    | Alunos | 16 | arrays de valores, `$avg`, `$unwind` |
    | Vendas | 17, 18, 19 | `$match`, `$group`, `$project`, `$multiply`, `$sum`, `$avg` |
    | Pedidos | 20 | array de documentos, `$unwind`, agrupamento por status |
    | Exercícios | todos | enunciado + comando no mongosh + resultado do banco |

    A rota **`/api/exercicios`** devolve, para cada item da lista, o enunciado, o comando
    equivalente no `mongosh` e o resultado real vindo do banco.
  contact:
    name: Repositório do trabalho
    url: http://localhost:3000
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Ambiente local
...
components:
  schemas:
    ObjectId:
      type: string
      pattern: ^[0-9a-fA-F]{24}$
      description: Identificador único do documento (24 caracteres hexadecimais).
      example: 6a98c80f3e8030cb5731e645
...
    ItemPedido:
      type: object
      required:
        - produto
        - quantidade
        - preco
      properties:
        produto:
          type: string
          example: Notebook
        quantidade:
          type: integer
          example: 1
        preco:
          type: number
          example: 4500
    ClientePedido:
      type: object
      properties:
        nome:
          type: string
          example: Ana Silva
        cidade:
          type: string
          example: Vitória da Conquista
    Pedido:
      type: object
      required:
        - numero
        - cliente
        - itens
      properties:
        _id:
          $ref: '#/components/schemas/ObjectId'
        numero:
          type: integer
          example: 1001
        cliente:
          $ref: '#/components/schemas/ClientePedido'
        itens:
          type: array
          items:
            $ref: '#/components/schemas/ItemPedido'
        status:
          $ref: '#/components/schemas/StatusPedido'
        formaPagamento:
          type: string
          example: Cartão
        valorTotal:
          type: number
          example: 4800
          description: Calculado pela API a partir dos itens.
        historicoStatus:
          type: array
          items:
            type: object
            properties:
              status:
                type: string
              data:
                type: string
                format: date-time
    StatusPedido:
      type: string
      enum:
        - Aguardando pagamento
        - Pago
        - Enviado
        - Entregue
        - Cancelado
      example: Pago
...
    ListaFiltrada:
      type: object
      description: 'Resposta padrão das consultas: repete o filtro aplicado e devolve os dados.'
      properties:
        filtro:
          type: object
          description: Filtro montado a partir da query string.
        total:
          type: integer
          example: 5
        dados:
          type: array
          items:
            type: object
...
    ResultadoRemocao:
      type: object
      properties:
        removidos:
          type: integer
          example: 1
...
    Erro:
      type: object
      properties:
        erro:
          type: string
          example: Recurso não encontrado.
  parameters:
...
    NumeroPedidoPath:
      name: numero
      in: path
      required: true
      description: Número do pedido (campo `numero`, não o `_id`).
      schema:
        type: integer
        example: 1001
  responses:
    RequisicaoInvalida:
      description: Dados inválidos ou campos obrigatórios ausentes.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    NaoEncontrado:
      description: Documento não encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
    Conflito:
      description: Conflito — o recurso já existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Erro'
paths:
...
  /api/pedidos:
    get:
      tags:
        - Pedidos
      summary: Lista pedidos com filtros (exercício 20, itens a a e)
      description: >
        `?cliente=` e `?cidade=` usam notação de ponto sobre o documento aninhado `cliente`; `?produto=` procura dentro
        do array de documentos `itens` (`itens.produto`).
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/StatusPedido'
          description: Exercício 20b.
        - name: cliente
          in: query
          schema:
            type: string
          example: Ana Silva
          description: Filtra por cliente.nome — exercício 20c.
        - name: cidade
          in: query
          schema:
            type: string
          example: Vitória da Conquista
          description: Filtra por cliente.cidade — exercício 20d.
        - name: produto
          in: query
          schema:
            type: string
          example: Notebook
          description: Filtra por itens.produto — exercício 20e.
        - name: formaPagamento
          in: query
          schema:
            type: string
            enum:
              - Cartão
              - Pix
              - Boleto
      responses:
        '200':
          description: Lista de pedidos ordenada por número.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ListaFiltrada'
                  - type: object
                    properties:
                      dados:
                        type: array
                        items:
                          $ref: '#/components/schemas/Pedido'
    post:
      tags:
        - Pedidos
      summary: Cria um pedido, calculando o valorTotal (exercício 20)
      description: >
        A API soma `quantidade × preco` de cada item e grava o resultado em `valorTotal`. O número do pedido é único —
        um número repetido devolve 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pedido'
            example:
              numero: 1007
              cliente:
                nome: Ana Silva
                cidade: Vitória da Conquista
              itens:
                - produto: Notebook
                  quantidade: 1
                  preco: 4500
                - produto: Mouse
                  quantidade: 2
                  preco: 150
              status: Pago
              formaPagamento: Cartão
      responses:
        '201':
          description: Pedido criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '409':
          $ref: '#/components/responses/Conflito'
...
  /api/pedidos/contagem:
    get:
      tags:
        - Pedidos
      summary: Conta pedidos, opcionalmente por status (exercício 20g)
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/StatusPedido'
      responses:
        '200':
          description: Total encontrado.
          content:
            application/json:
              schema:
                type: object
                properties:
                  filtro:
                    type: object
                  total:
                    type: integer
                    example: 2
...
  /api/pedidos/{numero}:
    get:
      tags:
        - Pedidos
      summary: Busca um pedido pelo número
      description: A busca usa o campo de negócio `numero`, e não o `_id` do MongoDB.
      parameters:
        - $ref: '#/components/parameters/NumeroPedidoPath'
      responses:
        '200':
          description: Pedido encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
    delete:
      tags:
        - Pedidos
      summary: Remove um pedido pelo número
      parameters:
        - $ref: '#/components/parameters/NumeroPedidoPath'
      responses:
        '200':
          description: Documento removido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultadoRemocao'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /api/pedidos/{numero}/status:
    patch:
      tags:
        - Pedidos
      summary: Altera o status do pedido (exercício 20f)
      description: >
        Aplica `$set` no status e `$push` em `historicoStatus`, registrando a data da mudança. Status fora da lista
        permitida devolve 400.
      parameters:
        - $ref: '#/components/parameters/NumeroPedidoPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  $ref: '#/components/schemas/StatusPedido'
            example:
              status: Enviado
      responses:
        '200':
          description: Pedido já atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
  /api/pedidos/{numero}/itens:
    post:
      tags:
        - Pedidos
      summary: Acrescenta um item ao pedido
      description: '`$push` no array `itens` e `$inc` no `valorTotal`, em uma única operação atômica.'
      parameters:
        - $ref: '#/components/parameters/NumeroPedidoPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemPedido'
            example:
              produto: Headset
              quantidade: 1
              preco: 300
      responses:
        '200':
          description: Pedido já atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          $ref: '#/components/responses/RequisicaoInvalida'
        '404':
          $ref: '#/components/responses/NaoEncontrado'
...